+++ CHANGELOG.md
@@ -12,6 +12,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `puffin` profiling to most example code - see [getting started guide](examples/getting-started.md)
   for more information
 
+### Changed
+
+- `ResolveMode` moved from `driver` to `driver::render_pass`
+
 ### Removed
 
 - Explicit reset of Vulkan command pools to reduce driver API calls